requirements.txt
asgiref==3.3.4
boto3==1.17.62
botocore==1.20.62
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
jmespath==0.10.0
Markdown==3.3.4
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
